HAR recording using webRequest.on* for the most part
But I don't think we actually need to write all this data to disk
after all even for a non-persistent background script (cf.
https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Background_scripts#record_state_changes),
since we can keep it in a local instead of a global variable in our
function, which should hopefully not be interrupted midway through
(which would be a problem anyway).

Committing this just in case.